+++ packages/core-components/src/lib/atoms/index.js
@@ -1,1 +1,2 @@
 export { default as Alert } from './alert/Alert.svelte';
+export * from './skeletons';