+++ packages/query-store/src/mutations/where.ts
@@ -1,0 +1,7 @@
+import { sql, Query } from '@uwdata/mosaic-sql';
+
+export const where = (q: Query, fragment: string): Query => {
+	q.where(sql`${fragment}`);
+
+	return q;
+};