+++ sites/example-project/sources/needful_things/item_by_category.sql
@@ -1,0 +1,6 @@
+select 
+    category,
+    item
+from orders
+group by item, category
+order by category, item